backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.0
pydantic-settings==2.3.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
email-validator==2.1.1
itsdangerous==2.1.2
APScheduler==3.10.4
bcrypt==4.0.1
